+++ rocketpy/__init__.py
@@ -1,5 +1,12 @@
+import logging
+
 from .control import _Controller
 from .environment import Environment, EnvironmentAnalysis
+from .exceptions import (
+    InvalidInertiaError,
+    InvalidParameterError,
+    UnstableRocketWarning,
+)
 from .mathutils import (
     Function,
     PiecewiseFunction,
@@ -18,6 +25,7 @@ from .motors import (
     MassFlowRateBasedTank,
     Motor,
     PointMassMotor,
+    RingClusterMotor,
     SolidMotor,
     SphericalTank,
     Tank,
@@ -48,6 +56,7 @@ from .rocket import (
     TrapezoidalFins,
 )
 from .sensitivity import SensitivityModel
+from . import utils
 from .sensors import Accelerometer, Barometer, GnssReceiver, Gyroscope
 from .simulation import Flight, MonteCarlo, MultivariateRejectionSampler
 from .stochastic import (
@@ -63,3 +72,5 @@ from .stochastic import (
     StochasticTail,
     StochasticTrapezoidalFins,
 )
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())