+++ rocketpy/plots/aero_surface_plots.py
@@ -3,6 +3,7 @@
 from abc import ABC, abstractmethod
 
 import matplotlib.pyplot as plt
+
 import numpy as np
 from matplotlib.patches import Ellipse