+++ rocketpy/plots/compare/compare_flights.py
@@ -1,10 +1,15 @@
 # TODO: remove this disable once the code is refactored
 # pylint: disable=nested-min-max
+import logging
+
 import matplotlib.pyplot as plt
 import numpy as np
 
 from ..plot_helpers import show_or_save_fig, show_or_save_plot
+
 from .compare import Compare
+
+logger = logging.getLogger(__name__)
 
 # TODO: needs to refactor this class to use the show_or_save_plot
 
@@ -94,7 +99,7 @@ class CompareFlights(Compare):  # pylint: disable=too-many-public-methods
         """
         show_or_save_fig(fig, filename)
         if filename:
-            print("Plot saved to file: " + filename)
+            logger.info("Plot saved to file: %s", filename)
         else:
             plt.show()
 
@@ -1084,7 +1089,7 @@ class CompareFlights(Compare):  # pylint: disable=too-many-public-methods
         None
         """
 
-        print("This method is not implemented yet")
+        logger.warning("This method is not implemented yet.")
 
     def attitude_frequency(
         self,
@@ -1123,7 +1128,7 @@ class CompareFlights(Compare):  # pylint: disable=too-many-public-methods
         None
         """
 
-        print("This method is not implemented yet")
+        logger.warning("This method is not implemented yet.")
 
     @staticmethod
     def compare_trajectories_3d(  # pylint: disable=too-many-statements